third_party/WebKit/LayoutTests/editing/selection/4895428-3-expected.txt

PASS document.activeElement is field
PASS successfullyParsed is true

TEST COMPLETE
